cleanup with display mode manager
